requirements.txt
streamlit>=1.30.0
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
faiss-cpu>=1.7.4
huggingface-hub>=0.20.0
Pillow>=10.0.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
xmltodict>=0.13.0
requests>=2.31.0
numpy>=1.24.0
wikipedia>=1.4.0
langgraph>=0.2.0
